Moved all solution related functions to solutionProvider.py
